# REF: Special case NumericIndex._append_same_dtype()