- [x] tests passed
- [x] passes `git diff master -u -- "*.py" | flake8 --diff`

Simple patch which results in a modest speedup (~8%) of the following:

```python
In [2]: idxes = [pd.Int64Index(range(i)) for i in range(1, 10)]*200
In [3]: %timeit idxes[0].append(idxes[1:])
```

I don't know whether this is worth the extra code... just inquiring so that I can finalize #16236.