@jreback 2017-08-22
looks fine
can u add some asv for concat of Index
ones that hit your various paths would be good
IOW concat RangeIndex, IntIndex and Index cases would be great