@toobaz 2017-08-22
> can u add some asv for concat of Index

``pd.concat()`` doesn't accept Index inputs, so I guess you mean ``Index.append()``: but then I think my previous PR already [addressed](https://github.com/pandas-dev/pandas/blob/2f00159da32c85c3b30b433f78a43e47677711a3/pandas/tests/indexes/test_range.py#L955) all the cases you mention. I'm pretty sure at least it touches the code paths I added in #16213, in #16236 and in this PR.