@jorisvandenbossche 2017-08-22
@toobaz indeed append, but question was about asv benchmark, not tests