@toobaz 2017-08-22
> @toobaz indeed append, but question was about asv benchmark, not tests

Aha, OK, I had no idea of what "asv" meant. Looking into it.