@toobaz 2017-08-23
I guess the benchmarking part is ready... but I wasn't able to test it (by the way, can you confirm that the content of [this page](https://github.com/pandas-dev/pandas/wiki/Performance-Testing) is obsolete?). Is it automatically ran anywhere?