@TomAugspurger 2017-08-23
@toobaz yeah, that page is out of date. New docs are [here](http://pandas-docs.github.io/pandas-docs-travis/contributing.html?highlight=asv#running-the-performance-test-suite) (I'm tempted to just remove that wiki page. Any objections?

Results are run nightly and posted to http://pandas.pydata.org/speed/pandas/ (still setting this up)