@jorisvandenbossche 2017-08-23
I removed the content of the page and put a link to the contributing docs for now.