@toobaz 2017-08-23
OK, the benchmarks now take ~5 ms. each.