@pep8speaks 2017-08-24
Hello @toobaz! Thanks for updating the PR.

 - In the file [`asv_bench/benchmarks/index_object.py`](https://github.com/pandas-dev/pandas/blob/6324b86604240ed80dbcfdd14fc95ca8dfd0ef9e/asv_bench/benchmarks/index_object.py), following are the PEP8 issues :

> [Line 209:34](https://github.com/pandas-dev/pandas/blob/6324b86604240ed80dbcfdd14fc95ca8dfd0ef9e/asv_bench/benchmarks/index_object.py#L209): [E226](https://duckduckgo.com/?q=pep8%20E226) missing whitespace around arithmetic operator