@toobaz 2017-08-24
> can can you post these results (commit before your recent fix of range index concat) to this commit

You mean here?

(notice I changed the benchmarks so the ``RangeIndex`` special-casing is actually tested - indexes are consecutive)

Before special-casing ``RangeIndex.append()`` (with 94ef7b6a ):

```python
In [19]: %timeit ridx[0].append(ridx[1:])
100 loops, best of 3: 10.3 ms per loop

In [20]: %timeit iidx[0].append(iidx[1:])
100 loops, best of 3: 9.5 ms per loop

In [21]: %timeit oidx[0].append(oidx[1:])
100 loops, best of 3: 10.6 ms per loop

```

Before special-casing ``NumericIndex._concat_same_dtype()`` (with 2f00159d ):

```python
In [6]: %timeit ridx[0].append(ridx[1:])
100 loops, best of 3: 6.15 ms per loop

In [7]: %timeit iidx[0].append(iidx[1:])
100 loops, best of 3: 10.9 ms per loop

In [8]: %timeit oidx[0].append(oidx[1:])
100 loops, best of 3: 11.7 ms per loop
```

Now (with 6324b866 ):
```python
In [6]: %timeit ridx[0].append(ridx[1:])
100 loops, best of 3: 6.08 ms per loop

In [7]: %timeit iidx[0].append(iidx[1:])
100 loops, best of 3: 9.71 ms per loop

In [8]: %timeit oidx[0].append(oidx[1:])
100 loops, best of 3: 11.7 ms per loop
```