@toobaz 2017-08-24
> @toobaz I wanted to see the asv results for these benchmarks;
> ```asv continuous master thisbranch --bench concat `` (you might want to name these index_concat actually)

OK, when I'll find the time to set up asv.

> and I don't care about the internal benchmarks (a _*) function, rather the public API .append()

Not following you here... indeed I'm testing only public methods.