@jreback 2017-08-25
asv_bench/benchmarks/index_concat.py is included but no changes.