@jorisvandenbossche 2017-08-25
> asv_bench/benchmarks/index_concat.py is included but no changes.

that's a leftover that should be removed, but the actual benchmarks are in the other file and time `append`