@toobaz 2017-08-26
@jreback @jorisvandenbossche 
Anyone has a hint of what's [happening](https://travis-ci.org/pandas-dev/pandas/jobs/268647210) in ``pandas.tests.indexes.test_category.TestCategoricalIndex.test_reindexing``? It fails randomly on Travis, (on [other branches](https://travis-ci.org/pandas-dev/pandas/jobs/268583991) too) and I really can't understand what could cause the error.