@jreback 2017-08-26
https://github.com/pandas-dev/pandas/issues/17323